Move ogr/ogrsf_frmts/s57/ddfrecordindex.cpp to frmts/iso8211
